requirements.txt
matplotlib==3.5.2
nltk==3.7
numpy==1.23.1
pandas==1.4.4
Pillow==9.2.0
preprocessor==1.1.3
Sastrawi==1.0.1
scikit_learn==1.1.2
scipy==1.9.1
seaborn==0.11.2
streamlit==1.11.0
torch==1.10.2
tqdm==4.64.1
transformers==4.18.0
tweepy==4.10.0
wordcloud==1.8.2.2